Basic support for large disks
This patch adds support for larger than 4GB (hard)disk images. There are two
limitations that are removed:
 - use 32-bit value for the filesize
 - use 32-bit value for the sector number
We will likely never hit the 2nd limitation, but changing this was little extra
effort.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13088 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/fdc/DMKDiskImage.cc
@@ -153,7 +153,7 @@ void DMKDiskImage::writeTrackImpl(byte track, byte side, const RawTrack& input)
 }
 
 
-void DMKDiskImage::readSectorImpl(unsigned logicalSector, byte* buf)
+void DMKDiskImage::readSectorImpl(size_t logicalSector, byte* buf)
 {
 	byte track, side, sector;
 	logToPhys(logicalSector, track, side, sector);
@@ -169,7 +169,7 @@ void DMKDiskImage::readSectorImpl(unsigned logicalSector, byte* buf)
 	rawTrack.readBlock(sectorInfo.dataIdx, 512, buf);
 }
 
-void DMKDiskImage::writeSectorImpl(unsigned logicalSector, const byte* buf)
+void DMKDiskImage::writeSectorImpl(size_t logicalSector, const byte* buf)
 {
 	byte track, side, sector;
 	logToPhys(logicalSector, track, side, sector);
@@ -186,7 +186,7 @@ void DMKDiskImage::writeSectorImpl(unsigned logicalSector, const byte* buf)
 	writeTrack(track, side, rawTrack);
 }
 
-unsigned DMKDiskImage::getNbSectorsImpl() const
+size_t DMKDiskImage::getNbSectorsImpl() const
 {
 	unsigned t = singleSided ? numTracks : (2 * numTracks);
 	return t * const_cast<DMKDiskImage*>(this)->getSectorsPerTrack();
