+++ src/fdc/DMKDiskImage.hh
@@ -24,9 +24,9 @@ public:
 	virtual void writeTrackImpl(byte track, byte side, const RawTrack& input);
 
 	// logical sector emulation for SectorAccessibleDisk
-	virtual void readSectorImpl(unsigned sector, byte* buf);
-	virtual void writeSectorImpl(unsigned sector, const byte* buf);
-	virtual unsigned getNbSectorsImpl() const;
+	virtual void readSectorImpl(size_t sector, byte* buf);
+	virtual void writeSectorImpl(size_t sector, const byte* buf);
+	virtual size_t getNbSectorsImpl() const;
 	virtual bool isWriteProtectedImpl() const;
 	virtual Sha1Sum getSha1Sum();
