+++ src/fdc/DSKDiskImage.cc
@@ -24,13 +24,13 @@ DSKDiskImage::~DSKDiskImage()
 {
 }
 
-void DSKDiskImage::readSectorImpl(unsigned sector, byte* buf)
+void DSKDiskImage::readSectorImpl(size_t sector, byte* buf)
 {
 	file->seek(sector * SECTOR_SIZE);
 	file->read(buf, SECTOR_SIZE);
 }
 
-void DSKDiskImage::writeSectorImpl(unsigned sector, const byte* buf)
+void DSKDiskImage::writeSectorImpl(size_t sector, const byte* buf)
 {
 	file->seek(sector * SECTOR_SIZE);
 	file->write(buf, SECTOR_SIZE);
