+++ src/fdc/DSKDiskImage.hh
@@ -18,8 +18,8 @@ public:
 	virtual ~DSKDiskImage();
 
 private:
-	virtual void readSectorImpl(unsigned sector, byte* buf);
-	virtual void writeSectorImpl(unsigned sector, const byte* buf);
+	virtual void readSectorImpl(size_t sector, byte* buf);
+	virtual void writeSectorImpl(size_t sector, const byte* buf);
 	virtual bool isWriteProtectedImpl() const;
 	virtual Sha1Sum getSha1Sum();
