+++ src/fdc/DiskPartition.hh
@@ -24,16 +24,16 @@ public:
 	              const std::shared_ptr<SectorAccessibleDisk>& owned = nullptr);
 
 	DiskPartition(SectorAccessibleDisk& parent,
-	              unsigned start, unsigned length);
+	              size_t start, size_t length);
 
 private:
-	virtual void readSectorImpl(unsigned sector, byte* buf);
-	virtual void writeSectorImpl(unsigned sector, const byte* buf);
+	virtual void readSectorImpl(size_t sector, byte* buf);
+	virtual void writeSectorImpl(size_t sector, const byte* buf);
 	virtual bool isWriteProtectedImpl() const;
 
 	SectorAccessibleDisk& parent;
 	std::shared_ptr<SectorAccessibleDisk> owned;
-	unsigned start;
+	size_t start;
 };
 
 } // namespace openmsx
