+++ src/fdc/DummyDisk.cc
@@ -21,12 +21,12 @@ bool DummyDisk::isWriteProtectedImpl() const
 	return true; // TODO check
 }
 
-void DummyDisk::readSectorImpl(unsigned /*sector*/, byte* /*buf*/)
+void DummyDisk::readSectorImpl(size_t /*sector*/, byte* /*buf*/)
 {
 	throw DriveEmptyException("No disk in drive");
 }
 
-void DummyDisk::writeSectorImpl(unsigned /*sector*/, const byte* /*buf*/)
+void DummyDisk::writeSectorImpl(size_t /*sector*/, const byte* /*buf*/)
 {
 	throw DriveEmptyException("No disk in drive");
 }
