+++ src/fdc/RamDSKDiskImage.hh
@@ -11,13 +11,13 @@ namespace openmsx {
 class RamDSKDiskImage : public SectorBasedDisk
 {
 public:
-	explicit RamDSKDiskImage(unsigned size = 720 * 1024);
+	explicit RamDSKDiskImage(size_t size = 720 * 1024);
 	virtual ~RamDSKDiskImage();
 
 private:
 	// SectorBasedDisk
-	virtual void readSectorImpl(unsigned sector, byte* buf);
-	virtual void writeSectorImpl(unsigned sector, const byte* buf);
+	virtual void readSectorImpl(size_t sector, byte* buf);
+	virtual void writeSectorImpl(size_t sector, const byte* buf);
 	virtual bool isWriteProtectedImpl() const;
 
 	MemBuffer<byte> diskdata;
