+++ src/fdc/XSADiskImage.hh
@@ -24,8 +24,8 @@ public:
 
 private:
 	// SectorBasedDisk
-	virtual void readSectorImpl(unsigned sector, byte* buf);
-	virtual void writeSectorImpl(unsigned sector, const byte* buf);
+	virtual void readSectorImpl(size_t sector, byte* buf);
+	virtual void writeSectorImpl(size_t sector, const byte* buf);
 	virtual bool isWriteProtectedImpl() const;
 
 	std::vector<byte> data;
