+++ src/ide/HD.hh
@@ -33,9 +33,9 @@ public:
 
 private:
 	// SectorAccessibleDisk:
-	virtual void readSectorImpl(unsigned sector, byte* buf);
-	virtual void writeSectorImpl(unsigned sector, const byte* buf);
-	virtual unsigned getNbSectorsImpl() const;
+	virtual void readSectorImpl(size_t sector, byte* buf);
+	virtual void writeSectorImpl(size_t sector, const byte* buf);
+	virtual size_t getNbSectorsImpl() const;
 	virtual bool isWriteProtectedImpl() const;
 	virtual Sha1Sum getSha1Sum();
 
@@ -53,7 +53,7 @@ private:
 
 	std::unique_ptr<File> file;
 	Filename filename;
-	unsigned filesize;
+	size_t filesize;
 	bool alreadyTried;
 };
