+++ src/ide/SCSILS120.hh
@@ -37,9 +37,9 @@ public:
 
 private:
 	// SectorAccessibleDisk:
-	virtual void readSectorImpl(unsigned sector, byte* buf);
-	virtual void writeSectorImpl(unsigned sector, const byte* buf);
-	virtual unsigned getNbSectorsImpl() const;
+	virtual void readSectorImpl(size_t sector, byte* buf);
+	virtual void writeSectorImpl(size_t sector, const byte* buf);
+	virtual size_t getNbSectorsImpl() const;
 	virtual bool isWriteProtectedImpl() const;
 	virtual Sha1Sum getSha1Sum();
